requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
streamlit-sortables>=0.2.0
gspread>=5.11.3
google-auth>=2.23.3